remove parallel executor unittest
