+++ tools/parallel_UT_rule.py
@@ -626,7 +626,6 @@ HIGH_PARALLEL_JOB_NEW = [
     'test_memory_analysis',
     'test_matrix_rank_op',
     'test_merged_momentum_op',
-    'test_parallel_executor_run_cinn',
     'test_parallel_dygraph_dataparallel_cpuonly',
     'test_eigvals_op',
     'test_sparse_attention_op',
@@ -672,9 +671,7 @@ SECONDARY_HIGH_PARALLEL_JOB_NEW = [
     'test_analyzer_int8_googlenet',
     'test_analyzer_seq_pool1_compare_determine',
     'save_quant2_model_ernie',
-    'test_parallel_executor_seresnext_with_fuse_all_reduce_cpu',
     'test_dataset_uci_housing',
-    'test_parallel_executor_seresnext_base_cpu',
     'test_dataset_download',
     'test_quant_int8_mobilenetv1_mkldnn',
     'test_crf_decoding_op',
@@ -690,7 +687,6 @@ SECONDARY_HIGH_PARALLEL_JOB_NEW = [
     'test_weight_quantization_mobilenetv1',
     'test_concat_mkldnn_op',
     'test_gaussian_random_mkldnn_op',
-    'test_parallel_executor_seresnext_with_reduce_cpu',
     'test_dataset_imikolov',
     'test_analyzer_rnn1',
     'test_conv2d_mkldnn_op',
@@ -809,7 +805,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_maximum_op',
     'test_rnn_cell_api',
     'device_code_test',
-    'test_ir_inplace_pass',
     'test_cos_sim_op',
     'test_lite_tensor_utils',
     'test_fit_a_line',
@@ -892,7 +887,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_scale_mkldnn_op',
     'test_load_state_dict_from_old_format',
     'test_lookup_table_v2_op',
-    'test_mix_precision_all_reduce_fuse',
     'test_spp_op',
     'test_op_converter',
     'test_mixed_vector',
@@ -923,7 +917,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_run_program_op',
     'test_cuda_random_seed',
     'test_linear_interp_op',
-    'test_fuse_all_reduce_pass',
     'tensor_util_test',
     'test_median',
     'test_nanmedian',
@@ -1029,7 +1022,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_gather_tree_op',
     'test_elementwise_mul_op',
     'test_cycle_gan',
-    'test_parallel_executor_transformer_auto_growth',
     'test_bitwise_op',
     'test_uniform_random_op',
     'trt_split_converter_test',
@@ -1086,7 +1078,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_imperative_layer_children',
     'nccl_op_test',
     'test_share_data_op',
-    'test_ir_memory_optimize_transformer',
     'test_math_op_patch',
     'test_base_layer',
     'test_dequantize_log_op',
@@ -1104,7 +1095,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_affine_channel_op',
     'test_leaky_relu_grad_grad_functor',
     'test_ctc_align',
-    'test_fuse_relu_depthwise_conv_pass',
     'test_complex_kron',
     'test_imperative_skip_op',
     'test_dgc_op',
@@ -1257,7 +1247,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_conv_elementwise_add2_act_fuse_pass',
     'test_imperative_container_layerlist',
     'test_dequantize_abs_max_op',
-    'test_fuse_optimizer_pass',
     'test_optimizer',
     'test_dynamic_rnn_stop_gradient',
     'test_raw_program_optimizer',
@@ -1359,7 +1348,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_gradient_accmulator',
     'test_instance_norm_op_v2',
     'test_mobile_net',
-    'test_parallel_executor_transformer',
     'test_tensor_scalar_type_promotion_dynamic',
     'test_eager_deletion_delete_vars',
     'test_asp_pruning_1d',
@@ -1386,7 +1374,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_tensorrt_engine',
     'test_affine_grid_function',
     'test_nonzero_api',
-    'test_ir_memory_optimize_pass',
     'test_reduce_mkldnn_op',
     'test_bilinear_interp_op',
     'test_cvm_op',
@@ -1468,9 +1455,6 @@ FOURTH_HIGH_PARALLEL_JOB_NEW = [
     'test_save_inference_model',
     'test_smooth_l1_loss',
     'test_bilateral_slice_op',
-    'test_parallel_executor_seresnext_base_gpu',
-    'test_parallel_executor_seresnext_with_fuse_all_reduce_gpu',
-    'test_parallel_executor_seresnext_with_reduce_gpu',
     'test_data_norm_op',
     'test_install_check',
     'graph_node_test',
@@ -2170,7 +2154,6 @@ CPU_PARALLEL_JOB = [
     'test_analyzer_capi_exp_xpu',
     'test_egr_task_autocodegen',
     'test_static_save_load_bf16',
-    'test_parallel_executor_run_cinn',
     'test_egr_task_tensor_utils',
     'test_egr_task_hook',
     'test_egr_task_forward_autograd',
@@ -2285,15 +2268,12 @@ TETRAD_PARALLEL_JOB = [
     'test_fused_transformer_encoder_layer',
     'test_eager_deletion_while_op',
     'test_dataloader_unkeep_order',
-    'test_parallel_executor_profiler',
     'test_correlation',
-    'test_ir_inplace_pass',
     'test_moving_average_abs_max_scale_op',
     'test_flatten_contiguous_range_op',
     'test_transforms',
     'test_sum_op',
     'test_scatter_op',
-    'test_mix_precision_all_reduce_fuse',
     'test_tensorrt_engine_op',
     'test_zeropad2d',
     'test_isclose_op',
@@ -2888,7 +2868,6 @@ TWO_PARALLEL_JOB = [
     'test_user_defined_quantization',
     'test_quantization_scale_pass',
     'feed_forward_test',
-    'test_fuse_optimizer_pass',
     'test_standalone_executor',
     'test_imperative_qat_user_defined',
     'test_mkldnn_fc_act_fuse_pass',
@@ -2896,7 +2875,6 @@ TWO_PARALLEL_JOB = [
     'test_signal',
     'test_fused_feedforward_op',
     'test_weight_decay_extend',
-    'test_fuse_relu_depthwise_conv_pass',
     'test_diag_v2',
     'test_tensordot',
     'test_rnn_decode_api',
@@ -2921,7 +2899,6 @@ TWO_PARALLEL_JOB = [
     'test_multinomial_op',
     'test_fused_elemwise_activation_op',
     'test_profiler',
-    'test_ir_memory_optimize_pass',
     'test_callback_reduce_lr_on_plateau',
     'test_paddle_save_load',
     'test_stack_op',
@@ -3063,10 +3040,8 @@ TWO_PARALLEL_JOB = [
     'test_squeeze2_mkldnn_op',
     'test_conv2d_transpose_bf16_mkldnn_op',
     'test_slice_mkldnn_op',
-    'test_parallel_executor_seresnext_base_cpu',
     'test_stack_mkldnn_op',
     'test_softplus_mkldnn_op',
-    'test_parallel_executor_seresnext_with_reduce_cpu',
     'test_nearest_interp_v2_mkldnn_op',
     'test_fusion_lstm_mkldnn_op',
     'test_fuse_resnet_unit',
@@ -3074,7 +3049,6 @@ TWO_PARALLEL_JOB = [
     'test_uniform_random_bf16_op',
     'test_reshape_mkldnn_op',
     'test_reduce_bf16_mkldnn_op',
-    'test_parallel_executor_seresnext_with_fuse_all_reduce_cpu',
     'test_nearest_interp_mkldnn_op',
     'test_ir_graph_to_program_pass',
     'test_fusion_lstm_int8_mkldnn_op',
