+++ tools/static_mode_white_list.py
@@ -217,9 +217,6 @@ STATIC_MODE_TESTING_LIST = [
     'test_functional_conv2d_transpose',
     'test_functional_conv3d',
     'test_functional_conv3d_transpose',
-    'test_fuse_all_reduce_pass',
-    'test_fuse_optimizer_pass',
-    'test_fuse_relu_depthwise_conv_pass',
     'test_fused_elemwise_activation_op',
     'test_fused_emb_seq_pool_op',
     'test_fused_embedding_fc_lstm_op',
@@ -280,7 +277,6 @@ STATIC_MODE_TESTING_LIST = [
     'test_inverse_op',
     'test_io_save_load',
     'test_iou_similarity_op',
-    'test_ir_memory_optimize_pass',
     'test_kldiv_loss_op',
     'test_kron_op',
     'test_l1_norm_op',
@@ -507,14 +503,8 @@ STATIC_MODE_TESTING_LIST = [
     'test_transpiler_ops',
     'test_communicator_sync',
     'test_collective_optimizer',
-    'test_parallel_executor_profiler',
-    'test_parallel_executor_transformer',
-    'test_parallel_executor_transformer_auto_growth',
     'test_data_norm_op',
     'test_fuse_bn_act_pass',
-    'test_parallel_executor_seresnext_base_cpu',
-    'test_parallel_executor_seresnext_with_reduce_cpu',
-    'test_parallel_executor_seresnext_with_fuse_all_reduce_cpu',
     'test_layers',
     'test_sequence_concat',
     'test_sequence_conv',
@@ -616,12 +606,9 @@ STATIC_MODE_TESTING_LIST = [
     'test_fleet_metric',
     'test_fused_bn_add_act',
     'test_fused_multihead_matmul_op',
-    'test_ir_inplace_pass',
-    'test_mix_precision_all_reduce_fuse',
     'test_rank_attention_op',
     'test_fleet_base',
     'test_fleet_meta_optimizer_base',
-    'test_ir_memory_optimize_transformer',
     'test_trt_fc_fuse_pass',
     'test_trt_quant_conv2d_dequant_fuse_pass',
     'test_trt_slice_plugin',
@@ -644,9 +631,6 @@ STATIC_MODE_TESTING_LIST = [
     'test_trt_pad_op',
     'test_trt_shuffle_channel_detect_pass',
     'test_trt_subgraph_pass',
-    'test_parallel_executor_seresnext_base_gpu',
-    'test_parallel_executor_seresnext_with_fuse_all_reduce_gpu',
-    'test_parallel_executor_seresnext_with_reduce_gpu',
     'test_sync_batch_norm_op',
     'test_multiprocess_dataloader_iterable_dataset_static',
     'test_multiprocess_dataloader_static',
